[App Design and Layout] Composing Complex Interfaces: Section 3 Create a Category Row
Landmarks displays each category in a row that scrolls horizontally. Add a new view type to represent the row, then display all the landmarks for that category in the new view.

Reuse parts of the Landmark view you created in [Creating and Combining Views](https://developer.apple.com/tutorials/swiftui/creating-and-combining-views) to create familiar previews of a landmark.