[Framework Integration] Interfacing with UIKit: Section 3 Track the Page in a SwiftUI View’s State
To prepare for adding a custom UIPageControl, you need a way to track the current page from within PageView.

To do this, you’ll declare a @State property in PageView, and pass a binding to this property down to the PageViewController view. The PageViewController updates the binding to match the visible page.